pyproject.toml
[tool.poetry]
name = "kms-sample"
version = "0.1.0"
description = ""
authors = [""]
readme = "README.md"
packages = [{include = "kms_sample"}]

[tool.poetry.dependencies]
python = "^3.10"
boto3 = "^1.26.140"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
